+++ src/renderer/plugins/Tips/Tips.js
@@ -2,7 +2,21 @@ import Tips from './Tips.vue'
 import Vue from 'vue'
 
 const TipsConstructor = Vue.extend(Tips)
-export default ({ position, message } = {}) => {
+
+const addAutoCloseTimer = (instance, time) => {
+  if (!time) return
+  if (instance.autoCloseTimer) clearTimeout(instance.autoCloseTimer)
+  instance.autoCloseTimer = setTimeout(() => {
+    instance.cancel()
+  }, time)
+}
+const clearAutoCloseTimer = instance => {
+  if (!instance.autoCloseTimer) return
+  clearTimeout(instance.autoCloseTimer)
+  instance.autoCloseTimer = null
+}
+
+export default ({ position, message, autoCloseTime } = {}) => {
   if (!position) return
   let instance = new TipsConstructor().$mount(document.createElement('div'))
 
@@ -17,13 +31,18 @@ export default ({ position, message } = {}) => {
   document.body.appendChild(instance.$el)
 
   instance.cancel = () => {
+    instance.$emit('beforeClose', instance)
+    clearAutoCloseTimer(instance)
     instance.visible = false
     instance = null
   }
 
   instance.setTips = tips => {
+    addAutoCloseTimer(instance, autoCloseTime)
     instance.message = tips
   }
+
+  addAutoCloseTimer(instance, autoCloseTime)
 
   return instance
 }